requirements.txt
psycopg2-binary
websocket_client
